moved matchObjectCollection3 def in RootTools/python/utils/DeltaR.py
