+++ CMGTools/TTHAnalysis/python/analyzers/ttHLepMCMatchAnalyzer.py
@@ -18,46 +18,6 @@ from CMGTools.RootTools.physicsobjects.Jet import Jet
 from CMGTools.RootTools.utils.DeltaR import *
 from CMGTools.RootTools.physicsobjects.genutils import *
 
-def matchObjectCollection3 ( objects, matchCollection, deltaRMax = 0.3, filter = lambda x,y : True ):
-    '''Univoque association of an element from matchCollection to an element of objects.
-    Reco and Gen objects get the "matched" attribute, true is they are re part of a matched tulpe.
-    By default, the matching is true only if delta R is smaller than 0.3.
-    '''
-    # 
-    pairs = {}
-    if len(objects)==0:
-            return pairs
-    if len(matchCollection)==0:
-            return dict( zip(objects, [None]*len(objects)) )
-    # build all possible combinations
-    allPairs = [(deltaR2 (object.eta(), object.phi(), match.eta(), match.phi()), (object, match)) for object in objects for match in matchCollection if filter(object,match) ]
-    allPairs.sort ()
-    #
-    # to flag already matched objects
-    # FIXME this variable remains appended to the object, I do not like it
-    for object in objects:
-        object.matched = False
-    for match in matchCollection:
-        match.matched = False
-    #
-    deltaR2Max = deltaRMax * deltaRMax
-    for dR2, (object, match) in allPairs:
-        if dR2 > deltaR2Max:
-            break
-        if dR2 < deltaR2Max and object.matched == False and match.matched == False:
-            object.matched = True
-            match.matched = True
-            pairs[object] = match
-    #
-    for object in objects:
-       if object.matched == False:
-           pairs[object] = None
-    #
-    return pairs
-    # by now, the matched attribute remains in the objects, for future usage
-    # one could remove it with delattr (object, attrname)
-
-        
 class ttHLepMCMatchAnalyzer( Analyzer ):
     def __init__(self, cfg_ana, cfg_comp, looperName ):
         super(ttHLepMCMatchAnalyzer,self).__init__(cfg_ana,cfg_comp,looperName)
